requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.1.0
python-dotenv==1.1.1
requests==2.32.3
torch==2.6.0
transformers==4.52.3
